Add files via upload
TASK-1 Prediction using Supervised ML
Simple Linear Regression
(LEVEL-BEGINNER)

Predicting the percentage of an student based on the no. of study hours.
This is a simple linear regression task as it involves just two variables.
Here, the Python Scikit-Learn library is used for machine learning to implement regression functions.
To Do: What will be predicted score if a student studies for 9.25 hrs/day.
Used Python language, in Jupyter Notebook.